refactor: Migrate hook domain logic from orbit-cli into orbit-core [ORB-00275]
Planned-By: codex
