+++ crates/orbit-cli/src/command/hook/render.rs
@@ -1,7 +1,4 @@
 use clap::ValueEnum;
-use orbit_common::types::LearningReminder;
-use orbit_core::OrbitError;
-use serde_json::json;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, ValueEnum)]
 pub enum HookOutputFormat {
@@ -11,42 +8,13 @@ pub enum HookOutputFormat {
     Grok,
 }
 
-pub fn render_reminders(
-    format: HookOutputFormat,
-    admitted: &[LearningReminder],
-) -> Result<String, OrbitError> {
-    match format {
-        HookOutputFormat::Claude | HookOutputFormat::Grok => Ok(render_claude(admitted)),
-        HookOutputFormat::Codex => render_codex(admitted),
-        HookOutputFormat::Gemini => render_gemini(admitted),
-    }
-}
-
-pub fn render_claude(admitted: &[LearningReminder]) -> String {
-    orbit_common::types::render_reminder_block(admitted)
-}
-
-pub fn render_codex(admitted: &[LearningReminder]) -> Result<String, OrbitError> {
-    render_json_context("PreToolUse", admitted)
-}
-
-pub fn render_gemini(admitted: &[LearningReminder]) -> Result<String, OrbitError> {
-    // Gemini CLI names its documented pre-tool hook event `BeforeTool`; the
-    // renderer stays separate so the wiring can change when Gemini's hook
-    // context surface settles.
-    render_json_context("BeforeTool", admitted)
-}
-
-fn render_json_context(
-    event_name: &str,
-    admitted: &[LearningReminder],
-) -> Result<String, OrbitError> {
-    let block = render_claude(admitted);
-    serde_json::to_string(&json!({
-        "hookSpecificOutput": {
-            "hookEventName": event_name,
-            "additionalContext": block,
+impl From<HookOutputFormat> for orbit_core::command::learning_hook::HookOutputFormat {
+    fn from(format: HookOutputFormat) -> Self {
+        match format {
+            HookOutputFormat::Claude => Self::Claude,
+            HookOutputFormat::Codex => Self::Codex,
+            HookOutputFormat::Gemini => Self::Gemini,
+            HookOutputFormat::Grok => Self::Grok,
         }
-    }))
-    .map_err(|error| OrbitError::Execution(format!("serialize hook output: {error}")))
+    }
 }
