+++ crates/orbit-core/src/command/learning_hook.rs
@@ -1,11 +1,21 @@
+use std::fs::{self, File, OpenOptions};
+use std::io::{Read, Seek, SeekFrom, Write};
 use std::path::{Path, PathBuf};
+use std::time::{Duration, Instant};
 
+use fs2::FileExt;
 use orbit_common::types::{
-    LearningInjectionCaps, LearningInjectionState, LearningReminder, OrbitError,
+    AuditEventStatus, LearningInjectionCaps, LearningInjectionState, LearningReminder, OrbitError,
 };
+use orbit_store::{AuditEventInsertParams, LearningSearchParams};
 use serde_json::Value;
+use serde_json::json;
 
 use crate::OrbitRuntime;
+use crate::redact_sensitive_env_text;
+
+const LOCK_RETRY_INTERVAL: Duration = Duration::from_millis(5);
+const LOCK_RETRY_BUDGET: Duration = Duration::from_millis(50);
 
 pub const ORBIT_BIN_ENV: &str = "ORBIT_BIN";
 pub const ORBIT_SESSION_ID_ENV: &str = "ORBIT_SESSION_ID";
@@ -18,6 +28,14 @@ pub type SessionLearningState = LearningInjectionState;
 pub struct HookPayload {
     pub tool_name: String,
     pub target_path: String,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum HookOutputFormat {
+    Claude,
+    Codex,
+    Gemini,
+    Grok,
 }
 
 pub const CLAUDE_PRETOOLUSE_TOOLS: &[&str] = &["Edit", "Write", "Read"];
@@ -33,6 +51,43 @@ pub const GEMINI_PRETOOLUSE_TOOLS: &[&str] = &[
     "Edit",
     "Bash",
 ];
+
+pub fn run_pretooluse(runtime: &OrbitRuntime, format: HookOutputFormat) -> Option<String> {
+    let start = Instant::now();
+    match run_pretooluse_payload(runtime, format, start) {
+        Ok(output) => output,
+        Err(error) => {
+            tracing::warn!(error = %redact_sensitive_env_text(&error), "learning hook failed open");
+            None
+        }
+    }
+}
+
+pub fn render_reminders(
+    format: HookOutputFormat,
+    admitted: &[LearningReminder],
+) -> Result<String, OrbitError> {
+    match format {
+        HookOutputFormat::Claude | HookOutputFormat::Grok => Ok(render_claude(admitted)),
+        HookOutputFormat::Codex => render_codex(admitted),
+        HookOutputFormat::Gemini => render_gemini(admitted),
+    }
+}
+
+pub fn render_claude(admitted: &[LearningReminder]) -> String {
+    orbit_common::types::render_reminder_block(admitted)
+}
+
+pub fn render_codex(admitted: &[LearningReminder]) -> Result<String, OrbitError> {
+    render_json_context("PreToolUse", admitted)
+}
+
+pub fn render_gemini(admitted: &[LearningReminder]) -> Result<String, OrbitError> {
+    // Gemini CLI names its documented pre-tool hook event `BeforeTool`; the
+    // renderer stays separate so the wiring can change when Gemini's hook
+    // context surface settles.
+    render_json_context("BeforeTool", admitted)
+}
 
 pub fn parse_payload(stdin: &str) -> Option<HookPayload> {
     parse_payload_with_tools(stdin, CLAUDE_PRETOOLUSE_TOOLS)
@@ -76,6 +131,236 @@ pub fn parse_payload_with_tools(stdin: &str, accepted_tools: &[&str]) -> Option<
         tool_name: tool_name.to_string(),
         target_path: target_path.to_string(),
     })
+}
+
+pub fn caps_from_env() -> LearningInjectionCaps {
+    LearningInjectionCaps {
+        per_call: cap_from_env(
+            ORBIT_LEARNING_PER_CALL_CAP_ENV,
+            orbit_common::types::DEFAULT_LEARNING_REMINDER_PER_CALL_CAP,
+        ),
+        per_session_hard: cap_from_env(
+            ORBIT_LEARNING_SESSION_CAP_ENV,
+            orbit_common::types::DEFAULT_LEARNING_REMINDER_SESSION_CAP,
+        ),
+    }
+}
+
+pub fn state_file_path(
+    repo_root: &Path,
+    session_id: Option<&str>,
+    tmpdir: &Path,
+    ppid: u32,
+) -> PathBuf {
+    match session_id.map(str::trim).filter(|value| !value.is_empty()) {
+        Some(session_id) => repo_root
+            .join(".orbit")
+            .join("state")
+            .join("sessions")
+            .join(session_id)
+            .join("learnings.json"),
+        None => tmpdir.join(format!("orbit-learning-hook-{ppid}.json")),
+    }
+}
+
+pub fn parse_state_json(raw: &str) -> SessionLearningState {
+    let Ok(value) = serde_json::from_str::<Value>(raw.trim()) else {
+        return SessionLearningState::new();
+    };
+    let Some(object) = value.as_object() else {
+        return SessionLearningState::new();
+    };
+    let emitted_ids = object
+        .get("emitted_ids")
+        .and_then(Value::as_array)
+        .into_iter()
+        .flatten()
+        .filter_map(|value| value.as_str().map(ToOwned::to_owned))
+        .collect::<std::collections::BTreeSet<_>>();
+    let count = object
+        .get("count")
+        .and_then(Value::as_u64)
+        .and_then(|value| usize::try_from(value).ok())
+        .unwrap_or(emitted_ids.len());
+    SessionLearningState { emitted_ids, count }
+}
+
+pub fn merge_state(
+    mut prior: SessionLearningState,
+    candidates: &[LearningReminder],
+    caps: LearningInjectionCaps,
+) -> (SessionLearningState, Vec<LearningReminder>) {
+    let admitted = prior.admit_reminders(candidates, caps);
+    (prior, admitted)
+}
+
+pub fn reminders_from_search_results(
+    results: Vec<orbit_store::LearningSearchResult>,
+) -> Vec<LearningReminder> {
+    results
+        .into_iter()
+        .map(|result| LearningReminder {
+            id: result.learning.id,
+            summary: result.learning.summary,
+            comments: Vec::new(),
+        })
+        .collect()
+}
+
+pub(crate) fn update_state_file(
+    state_path: &Path,
+    candidates: &[LearningReminder],
+    caps: LearningInjectionCaps,
+) -> Result<Vec<LearningReminder>, String> {
+    if let Some(parent) = state_path.parent() {
+        fs::create_dir_all(parent)
+            .map_err(|error| format!("create state dir {}: {error}", parent.display()))?;
+    }
+
+    let mut file = OpenOptions::new()
+        .read(true)
+        .write(true)
+        .create(true)
+        .truncate(false)
+        .open(state_path)
+        .map_err(|error| format!("open state file {}: {error}", state_path.display()))?;
+
+    try_lock_exclusive(&file)?;
+    let update_result = update_locked_state(&mut file, candidates, caps);
+    let unlock_result = file
+        .unlock()
+        .map_err(|error| format!("unlock state file {}: {error}", state_path.display()));
+    match (update_result, unlock_result) {
+        (Ok(admitted), Ok(())) => Ok(admitted),
+        (Err(error), _) => Err(error),
+        (Ok(_), Err(error)) => Err(error),
+    }
+}
+
+impl OrbitRuntime {
+    pub fn learning_hook_target_is_searchable(
+        &self,
+        target_path: &str,
+    ) -> Result<bool, OrbitError> {
+        let normalized = target_path.trim().replace('\\', "/");
+        if normalized == "~" || normalized.starts_with("~/") {
+            return Ok(false);
+        }
+
+        crate::command::learning::learning_search_path_matches_workspace(
+            &self.paths().repo_root,
+            target_path,
+        )
+    }
+
+    pub fn learning_hook_state_file_path(
+        &self,
+        session_id: Option<&str>,
+        tmpdir: &Path,
+        ppid: u32,
+    ) -> PathBuf {
+        state_file_path(&self.paths().repo_root, session_id, tmpdir, ppid)
+    }
+}
+
+fn run_pretooluse_payload(
+    runtime: &OrbitRuntime,
+    format: HookOutputFormat,
+    start: Instant,
+) -> Result<Option<String>, String> {
+    let mut stdin = String::new();
+    std::io::stdin()
+        .read_to_string(&mut stdin)
+        .map_err(|error| format!("read stdin: {error}"))?;
+
+    run_pretooluse_input(runtime, &stdin, format, start)
+}
+
+pub(crate) fn run_pretooluse_input(
+    runtime: &OrbitRuntime,
+    stdin: &str,
+    format: HookOutputFormat,
+    start: Instant,
+) -> Result<Option<String>, String> {
+    let Some(payload) = parse_payload_with_tools(stdin, accepted_tools(format)) else {
+        return Ok(None);
+    };
+
+    let caps = caps_from_env();
+    if !runtime
+        .learning_hook_target_is_searchable(&payload.target_path)
+        .map_err(|error| format!("classify learning target path: {error}"))?
+    {
+        return Ok(None);
+    }
+
+    let results = runtime
+        .search_learnings(LearningSearchParams {
+            path: Some(payload.target_path.clone()),
+            tag: None,
+            query: None,
+            limit: Some(caps.per_call),
+        })
+        .map_err(|error| format!("search learnings: {error}"))?;
+    if results.is_empty() {
+        return Ok(None);
+    }
+
+    let candidates = reminders_from_search_results(results);
+    let session_id = std::env::var(ORBIT_SESSION_ID_ENV)
+        .ok()
+        .filter(|value| !value.trim().is_empty());
+    let tmpdir = learning_hook_tmpdir();
+    let state_path =
+        runtime.learning_hook_state_file_path(session_id.as_deref(), &tmpdir, parent_process_id());
+    let admitted = update_state_file(&state_path, &candidates, caps)?;
+    if admitted.is_empty() {
+        return Ok(None);
+    }
+
+    emit_learning_injected_audit(
+        runtime,
+        &payload.tool_name,
+        &payload.target_path,
+        session_id.as_deref(),
+        &admitted,
+        start.elapsed(),
+    )?;
+
+    render_reminders(format, &admitted)
+        .map(Some)
+        .map_err(|error| format!("render reminders: {error}"))
+}
+
+pub(crate) fn accepted_tools(format: HookOutputFormat) -> &'static [&'static str] {
+    match format {
+        HookOutputFormat::Claude | HookOutputFormat::Grok => CLAUDE_PRETOOLUSE_TOOLS,
+        HookOutputFormat::Codex => CODEX_PRETOOLUSE_TOOLS,
+        HookOutputFormat::Gemini => GEMINI_PRETOOLUSE_TOOLS,
+    }
+}
+
+pub(crate) fn learning_hook_tmpdir() -> PathBuf {
+    std::env::var("TMPDIR")
+        .ok()
+        .map(|value| value.trim().to_string())
+        .filter(|value| !value.is_empty())
+        .map(PathBuf::from)
+        .unwrap_or_else(|| PathBuf::from("/tmp"))
+}
+
+fn render_json_context(
+    event_name: &str,
+    admitted: &[LearningReminder],
+) -> Result<String, OrbitError> {
+    let block = render_claude(admitted);
+    serde_json::to_string(&json!({
+        "hookSpecificOutput": {
+            "hookEventName": event_name,
+            "additionalContext": block,
+        }
+    }))
+    .map_err(|error| OrbitError::Execution(format!("serialize hook output: {error}")))
 }
 
 fn tool_name_allowed(tool_name: &str, accepted_tools: &[&str]) -> bool {
@@ -184,19 +469,6 @@ fn looks_like_path(token: &str) -> bool {
         .any(|suffix| token.ends_with(suffix))
 }
 
-pub fn caps_from_env() -> LearningInjectionCaps {
-    LearningInjectionCaps {
-        per_call: cap_from_env(
-            ORBIT_LEARNING_PER_CALL_CAP_ENV,
-            orbit_common::types::DEFAULT_LEARNING_REMINDER_PER_CALL_CAP,
-        ),
-        per_session_hard: cap_from_env(
-            ORBIT_LEARNING_SESSION_CAP_ENV,
-            orbit_common::types::DEFAULT_LEARNING_REMINDER_SESSION_CAP,
-        ),
-    }
-}
-
 fn cap_from_env(name: &str, default: usize) -> usize {
     std::env::var(name)
         .ok()
@@ -205,318 +477,113 @@ fn cap_from_env(name: &str, default: usize) -> usize {
         .unwrap_or(default)
 }
 
-pub fn state_file_path(
-    repo_root: &Path,
-    session_id: Option<&str>,
-    tmpdir: &Path,
-    ppid: u32,
-) -> PathBuf {
-    match session_id.map(str::trim).filter(|value| !value.is_empty()) {
-        Some(session_id) => repo_root
-            .join(".orbit")
-            .join("state")
-            .join("sessions")
-            .join(session_id)
-            .join("learnings.json"),
-        None => tmpdir.join(format!("orbit-learning-hook-{ppid}.json")),
+fn try_lock_exclusive(file: &File) -> Result<(), String> {
+    let started = Instant::now();
+    loop {
+        match file.try_lock_exclusive() {
+            Ok(()) => return Ok(()),
+            Err(error) if error.kind() == std::io::ErrorKind::WouldBlock => {
+                if started.elapsed() >= LOCK_RETRY_BUDGET {
+                    return Err("state file lock timed out".to_string());
+                }
+                std::thread::sleep(LOCK_RETRY_INTERVAL);
+            }
+            Err(error) => return Err(format!("lock state file: {error}")),
+        }
     }
 }
 
-pub fn parse_state_json(raw: &str) -> SessionLearningState {
-    let Ok(value) = serde_json::from_str::<Value>(raw.trim()) else {
-        return SessionLearningState::new();
-    };
-    let Some(object) = value.as_object() else {
-        return SessionLearningState::new();
-    };
-    let emitted_ids = object
-        .get("emitted_ids")
-        .and_then(Value::as_array)
-        .into_iter()
-        .flatten()
-        .filter_map(|value| value.as_str().map(ToOwned::to_owned))
-        .collect::<std::collections::BTreeSet<_>>();
-    let count = object
-        .get("count")
-        .and_then(Value::as_u64)
-        .and_then(|value| usize::try_from(value).ok())
-        .unwrap_or(emitted_ids.len());
-    SessionLearningState { emitted_ids, count }
-}
-
-pub fn merge_state(
-    mut prior: SessionLearningState,
+fn update_locked_state(
+    file: &mut File,
     candidates: &[LearningReminder],
     caps: LearningInjectionCaps,
-) -> (SessionLearningState, Vec<LearningReminder>) {
-    let admitted = prior.admit_reminders(candidates, caps);
-    (prior, admitted)
+) -> Result<Vec<LearningReminder>, String> {
+    file.seek(SeekFrom::Start(0))
+        .map_err(|error| format!("seek state file: {error}"))?;
+    let mut raw = String::new();
+    file.read_to_string(&mut raw)
+        .map_err(|error| format!("read state file: {error}"))?;
+
+    let prior = parse_state_json(&raw);
+    let (next_state, admitted) = merge_state(prior, candidates, caps);
+
+    file.seek(SeekFrom::Start(0))
+        .map_err(|error| format!("rewind state file: {error}"))?;
+    file.set_len(0)
+        .map_err(|error| format!("truncate state file: {error}"))?;
+    serde_json::to_writer_pretty(&mut *file, &next_state)
+        .map_err(|error| format!("serialize state file: {error}"))?;
+    file.write_all(b"\n")
+        .map_err(|error| format!("write state file: {error}"))?;
+    file.flush()
+        .map_err(|error| format!("flush state file: {error}"))?;
+
+    Ok(admitted)
 }
 
-pub fn reminders_from_search_results(
-    results: Vec<orbit_store::LearningSearchResult>,
-) -> Vec<LearningReminder> {
-    results
-        .into_iter()
-        .map(|result| LearningReminder {
-            id: result.learning.id,
-            summary: result.learning.summary,
-            comments: Vec::new(),
-        })
-        .collect()
+fn emit_learning_injected_audit(
+    runtime: &OrbitRuntime,
+    tool_name: &str,
+    target_path: &str,
+    session_id: Option<&str>,
+    admitted: &[LearningReminder],
+    duration: Duration,
+) -> Result<(), String> {
+    let learning_ids = admitted
+        .iter()
+        .map(|reminder| reminder.id.clone())
+        .collect::<Vec<_>>();
+    let arguments_json = serde_json::to_string(&json!({ "learning_ids": learning_ids }))
+        .map_err(|error| format!("serialize audit arguments: {error}"))?;
+    let working_directory = std::env::current_dir()
+        .map(|path| path.to_string_lossy().to_string())
+        .unwrap_or_else(|_| ".".to_string());
+
+    let params = AuditEventInsertParams {
+        execution_id: orbit_common::types::audit_execution_id("learning"),
+        command: "hook".to_string(),
+        subcommand: Some("pretooluse".to_string()),
+        tool_name: Some(tool_name.to_string()),
+        target_type: Some("learning_injected".to_string()),
+        target_id: Some(target_path.to_string()),
+        role: "hook".to_string(),
+        status: AuditEventStatus::Success,
+        exit_code: 0,
+        duration_ms: duration.as_millis() as i64,
+        working_directory,
+        arguments_json: Some(arguments_json),
+        stdout_truncated: None,
+        stderr_truncated: None,
+        error_message: None,
+        host: std::env::var("HOSTNAME").ok(),
+        pid: std::process::id(),
+        session_id: session_id.map(ToOwned::to_owned),
+        task_id: std::env::var("ORBIT_TASK_ID")
+            .ok()
+            .filter(|value| !value.is_empty()),
+        job_run_id: std::env::var("ORBIT_RUN_ID")
+            .ok()
+            .filter(|value| !value.is_empty()),
+        activity_id: std::env::var("ORBIT_ACTIVITY_ID")
+            .ok()
+            .filter(|value| !value.is_empty()),
+        step_index: std::env::var("ORBIT_STEP_INDEX")
+            .ok()
+            .and_then(|value| value.parse().ok()),
+    };
+
+    runtime
+        .record_audit_event(&params)
+        .map_err(|error| format!("record learning audit event: {error}"))
 }
 
-impl OrbitRuntime {
-    pub fn learning_hook_target_is_searchable(
-        &self,
-        target_path: &str,
-    ) -> Result<bool, OrbitError> {
-        let normalized = target_path.trim().replace('\\', "/");
-        if normalized == "~" || normalized.starts_with("~/") {
-            return Ok(false);
-        }
-
-        crate::command::learning::learning_search_path_matches_workspace(
-            &self.paths().repo_root,
-            target_path,
-        )
-    }
-
-    pub fn learning_hook_state_file_path(
-        &self,
-        session_id: Option<&str>,
-        tmpdir: &Path,
-        ppid: u32,
-    ) -> PathBuf {
-        state_file_path(&self.paths().repo_root, session_id, tmpdir, ppid)
-    }
+#[cfg(unix)]
+fn parent_process_id() -> u32 {
+    // SAFETY: getppid has no preconditions and only reads process metadata.
+    unsafe { libc::getppid() as u32 }
 }
 
-#[cfg(test)]
-mod tests {
-    use std::sync::{Mutex, MutexGuard, OnceLock};
-
-    use orbit_common::types::{LearningInjectionCaps, LearningReminder};
-
-    use super::*;
-
-    #[test]
-    fn parse_payload_accepts_tool_and_path_variants() {
-        let nested =
-            parse_payload(r#"{"tool_name":"Edit","tool_input":{"file_path":" src/lib.rs "}}"#)
-                .expect("nested payload");
-        assert_eq!(nested.tool_name, "Edit");
-        assert_eq!(nested.target_path, "src/lib.rs");
-
-        let camel = parse_payload(r#"{"toolName":"Write","toolInput":{"filePath":"README.md"}}"#)
-            .expect("camel payload");
-        assert_eq!(camel.tool_name, "Write");
-        assert_eq!(camel.target_path, "README.md");
-
-        let top_level = parse_payload(r#"{"tool_name":"Read","path":"Cargo.toml"}"#)
-            .expect("top-level payload");
-        assert_eq!(top_level.tool_name, "Read");
-        assert_eq!(top_level.target_path, "Cargo.toml");
-    }
-
-    #[test]
-    fn parse_payload_rejects_malformed_irrelevant_or_pathless_payloads() {
-        assert!(parse_payload("").is_none());
-        assert!(parse_payload("not-json").is_none());
-        assert!(parse_payload(r#"{"tool_name":"Bash","path":"src/lib.rs"}"#).is_none());
-        assert!(parse_payload(r#"{"tool_name":"Edit","path":"   "}"#).is_none());
-        assert!(parse_payload(r#"{"tool_name":"Edit"}"#).is_none());
-    }
-
-    #[test]
-    fn parse_payload_with_tools_accepts_codex_path_shapes() {
-        let bash = parse_payload_with_tools(
-            r#"{"tool_name":"Bash","tool_input":{"command":"sed -n '1,20p' crates/orbit-core/src/lib.rs"}}"#,
-            CODEX_PRETOOLUSE_TOOLS,
-        )
-        .expect("bash payload");
-        assert_eq!(bash.tool_name, "Bash");
-        assert_eq!(bash.target_path, "crates/orbit-core/src/lib.rs");
-
-        let patch = parse_payload_with_tools(
-            r#"{"tool_name":"apply_patch","tool_input":{"patch":"*** Begin Patch\n*** Update File: crates/orbit-cli/src/main.rs\n@@\n*** End Patch\n"}}"#,
-            CODEX_PRETOOLUSE_TOOLS,
-        )
-        .expect("patch payload");
-        assert_eq!(patch.tool_name, "apply_patch");
-        assert_eq!(patch.target_path, "crates/orbit-cli/src/main.rs");
-
-        let mcp = parse_payload_with_tools(
-            r#"{"tool_name":"mcp__plugin_orbit__fs_read","tool_input":{"filePaths":["README.md","Cargo.toml"]}}"#,
-            CODEX_PRETOOLUSE_TOOLS,
-        )
-        .expect("mcp payload");
-        assert_eq!(mcp.target_path, "README.md");
-    }
-
-    #[test]
-    fn state_file_path_matches_session_and_tmp_layouts() {
-        let repo_root = Path::new("/repo");
-        let tmpdir = Path::new("/tmp");
-        assert_eq!(
-            state_file_path(repo_root, Some("session-1"), tmpdir, 123),
-            PathBuf::from("/repo/.orbit/state/sessions/session-1/learnings.json")
-        );
-        assert_eq!(
-            state_file_path(repo_root, None, tmpdir, 123),
-            PathBuf::from("/tmp/orbit-learning-hook-123.json")
-        );
-    }
-
-    #[test]
-    fn parse_state_json_defaults_malformed_or_missing_count() {
-        assert_eq!(parse_state_json("not-json"), SessionLearningState::new());
-        let state = parse_state_json(r#"{"emitted_ids":["L2","L1"]}"#);
-        assert_eq!(state.count, 2);
-        assert!(state.emitted_ids.contains("L1"));
-        assert!(state.emitted_ids.contains("L2"));
-    }
-
-    #[test]
-    fn merge_state_admits_cold_candidates_and_dedups_warm_state() {
-        let candidates = reminders(&["L1", "L2"]);
-        let caps = LearningInjectionCaps {
-            per_call: 5,
-            per_session_hard: 20,
-        };
-
-        let (state, admitted) = merge_state(SessionLearningState::new(), &candidates, caps);
-        assert_eq!(
-            admitted.iter().map(|r| r.id.as_str()).collect::<Vec<_>>(),
-            ["L1", "L2"]
-        );
-        assert_eq!(state.count, 2);
-
-        let (_state, admitted) = merge_state(state, &candidates, caps);
-        assert!(admitted.is_empty());
-    }
-
-    #[test]
-    fn merge_state_honors_per_call_and_session_caps() {
-        let candidates = reminders(&["L1", "L2", "L3", "L4"]);
-        let (state, admitted) = merge_state(
-            SessionLearningState::new(),
-            &candidates,
-            LearningInjectionCaps {
-                per_call: 2,
-                per_session_hard: 20,
-            },
-        );
-        assert_eq!(
-            admitted.iter().map(|r| r.id.as_str()).collect::<Vec<_>>(),
-            ["L1", "L2"]
-        );
-        assert_eq!(state.count, 2);
-
-        let seeded = SessionLearningState::seeded(["L1".to_string(), "L2".to_string()]);
-        let (state, admitted) = merge_state(
-            seeded,
-            &candidates,
-            LearningInjectionCaps {
-                per_call: 5,
-                per_session_hard: 2,
-            },
-        );
-        assert!(admitted.is_empty());
-        assert_eq!(state.count, 2);
-    }
-
-    #[test]
-    fn caps_from_env_uses_shell_compatible_defaults_and_minimums() {
-        let _guard = EnvGuard::set(&[
-            (ORBIT_LEARNING_PER_CALL_CAP_ENV, None),
-            (ORBIT_LEARNING_SESSION_CAP_ENV, None),
-        ]);
-        assert_eq!(
-            caps_from_env(),
-            LearningInjectionCaps {
-                per_call: 5,
-                per_session_hard: 20,
-            }
-        );
-        drop(_guard);
-
-        let _guard = EnvGuard::set(&[
-            (ORBIT_LEARNING_PER_CALL_CAP_ENV, Some("2")),
-            (ORBIT_LEARNING_SESSION_CAP_ENV, Some("0")),
-        ]);
-        assert_eq!(
-            caps_from_env(),
-            LearningInjectionCaps {
-                per_call: 2,
-                per_session_hard: 1,
-            }
-        );
-        drop(_guard);
-
-        let _guard = EnvGuard::set(&[
-            (ORBIT_LEARNING_PER_CALL_CAP_ENV, Some("nope")),
-            (ORBIT_LEARNING_SESSION_CAP_ENV, Some("also-nope")),
-        ]);
-        assert_eq!(
-            caps_from_env(),
-            LearningInjectionCaps {
-                per_call: 5,
-                per_session_hard: 20,
-            }
-        );
-    }
-
-    fn reminders(ids: &[&str]) -> Vec<LearningReminder> {
-        ids.iter()
-            .map(|id| LearningReminder {
-                id: (*id).to_string(),
-                summary: format!("summary {id}"),
-                comments: Vec::new(),
-            })
-            .collect()
-    }
-
-    struct EnvGuard {
-        _lock: MutexGuard<'static, ()>,
-        saved: Vec<(&'static str, Option<String>)>,
-    }
-
-    impl EnvGuard {
-        fn set(values: &[(&'static str, Option<&str>)]) -> Self {
-            static LOCK: OnceLock<Mutex<()>> = OnceLock::new();
-            let lock = LOCK
-                .get_or_init(|| Mutex::new(()))
-                .lock()
-                .unwrap_or_else(|poisoned| poisoned.into_inner());
-            let saved = values
-                .iter()
-                .map(|(name, _)| (*name, std::env::var(name).ok()))
-                .collect::<Vec<_>>();
-            for (name, value) in values {
-                // SAFETY: EnvGuard serializes these process-wide mutations and restores them on drop.
-                unsafe {
-                    match value {
-                        Some(value) => std::env::set_var(name, value),
-                        None => std::env::remove_var(name),
-                    }
-                }
-            }
-            Self { _lock: lock, saved }
-        }
-    }
-
-    impl Drop for EnvGuard {
-        fn drop(&mut self) {
-            for (name, value) in &self.saved {
-                // SAFETY: EnvGuard holds the serialization lock until all saved values are restored.
-                unsafe {
-                    match value {
-                        Some(value) => std::env::set_var(name, value),
-                        None => std::env::remove_var(name),
-                    }
-                }
-            }
-        }
-    }
+#[cfg(not(unix))]
+fn parent_process_id() -> u32 {
+    std::process::id()
 }
