+++ crates/orbit-core/src/command/tests/mod.rs
@@ -1,0 +1,2 @@
+mod hook_install;
+mod learning_hook;
